+++ MCMaps/Models/Tips/PinPhotoOnboardingTip.swift
@@ -1,0 +1,27 @@
+//
+//  PinPhotoOnboardingTip.swift
+//  MCMaps
+//
+//  Created by Marquis Kurt on 22-03-2025.
+//
+
+import SwiftUI
+import TipKit
+
+struct PinPhotoOnboardingTip: Tip {
+    var options: [Option] {
+        [MaxDisplayCount(3)]
+    }
+
+    var title: Text {
+        Text("Add a photo")
+    }
+
+    var message: Text? {
+        Text("Add photos of your world to make this pin memorable.")
+    }
+
+    var image: Image? {
+        Image(systemName: "photo.badge.plus")
+    }
+}